perf: disable auto install duckdb vss extension 1. will cause macOS sign problem 2. will increase 40Mb for build
